+++ src/index.js
@@ -1,0 +1,2 @@
+import './js/mobile-menu'
+import './js/section-services'
